views.py v1

+++ golos/views.py
@@ -1,4 +1,5 @@
 from .models import *
+from django.contrib.auth.models import User
 try:
     from django.utils import simplejson as json
 except ImportError:
@@ -23,36 +24,38 @@ class VotesView(View):
     model = Nominees  # Модель данных - Статьи или Комментарии
     vote_type = None
 
-    def post(self, request, pk, key_b):
+    def post(self, request, pk):
         obj = Nominees.objects.get(pk=pk)
-        # GenericForeignKey не поддерживает метод get_or_create
-        try:
-            if key_b not in UserGolos:
-                key_b = UserGolos.objects.create(key=key_b)
-                key_b.save()
-            key_b = UserGolos.objects.get(key=key_b)
-            likedislike = LikeDislike.objects.get(content_type=ContentType.objects.get_for_model(obj), object_id=obj.id,
-                                                  user=key_b)
-            if likedislike.vote is not self.vote_type:
-                likedislike.vote = self.vote_type
-                likedislike.save(update_fields=['vote'])
-                result = True
-            else:
-                likedislike.delete()
-                result = False
-        except LikeDislike.DoesNotExist:
-            obj.votes.create(user=request.user, vote=self.vote_type)
-            result = True
 
-        return HttpResponse(
-            json.dumps({
-                "result": result,
-                "like_count": obj.votes.likes().count(),
-                "dislike_count": obj.votes.dislikes().count(),
-                "sum_rating": obj.votes.sum_rating()
-            }),
-            content_type="application/json"
-        )
+        # GenericForignKey не поддерживает метод get_or_create
+        if request.method == "POST":
+            kwargs = request.POST
+            if UserGolos.objects.all().filter(key=kwargs['key_b']).count() == 0:
+                key_b = UserGolos.objects.create(key=kwargs['key_b'])
+                key_b.save()
+            try:
+                key = UserGolos.objects.get(key=kwargs['key_b'])
+                likedislike = LikeDislike.objects.get(content_type=ContentType.objects.get_for_model(obj), object_id=obj.id, user=key)
+                if likedislike.vote is not self.vote_type:
+                    likedislike.vote = self.vote_type
+                    likedislike.save(update_fields=['vote'])
+                    result = True
+                else:
+                    likedislike.delete()
+                    result = False
+            except LikeDislike.DoesNotExist:
+                obj.votes.create(user=request.user, vote=self.vote_type)
+                result = True
+
+            return HttpResponse(
+                json.dumps({
+                    "result": result,
+                    "like_count": obj.votes.likes().count(),
+                    "dislike_count": obj.votes.dislikes().count(),
+                    "sum_rating": obj.votes.sum_rating()
+                }),
+                content_type="application/json"
+            )
 
     # def index(request):
 #     q = Show.objects.all()
